pyproject.toml
[tool.poetry]
name = "online-store"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "online_store"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
pillow = "^10.1.0"
django-ckeditor = "^6.7.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
